OperatingSystems
Operating Systems is the course I had in the fourth year of my study and
this code is one of the assignments we had during the course. It deals
with testing threading capabilities of operating system and how they
impact the performance of the system. Furthermore, whe do we consider
threading a bottleneck and when it is a good thing